promisesForPopulate moved to utils. Logout returns promise.
